fixed the cgmms

+++ Appretto/src/reader.c
@@ -41,12 +41,11 @@ int read_binary_blob(char *data_out,char *path,const char *expected_record,int m
       if(rank==0 && debug>1) printf("found record: %s\n",header);
       if(strcmp(expected_record,header)==0)
 	{
-	  int nbytes=lemonReaderBytes(reader);
+	  uint64_t nbytes=lemonReaderBytes(reader);
 	  nbytes_per_site=nbytes/glb_vol;
 	  if(nbytes_per_site>max_nbytes_per_site && rank==0)
 	    {
-	      fprintf(stderr,"Opsss! The file contain %d bytes per site and it is supposed to contain not more than: %d or %d\n",
-		      nbytes,nbytes_per_site,max_nbytes_per_site);
+	      fprintf(stderr,"Opsss! The file contain %d bytes per site and it is supposed to contain not more than: %d or %d\n",nbytes,nbytes_per_site,max_nbytes_per_site);
 	      fflush(stderr);
 	      MPI_Abort(MPI_COMM_WORLD,1);
 	    }
@@ -65,7 +64,7 @@ int read_binary_blob(char *data_out,char *path,const char *expected_record,int m
 	    {
 	      MPI_Barrier(cart_comm);
 	      double tac1=MPI_Wtime();
-	      if(rank==0) printf("Time elapsed by lemon to read %d bytes: %f s\n",nbytes,tac1-tic1);
+	      if(rank==0) printf("Time elapsed by lemon to read %ld bytes: %f s\n",nbytes,tac1-tic1);
 	    }
 
 	  read=1;
